functionalizing: moved constant tensors into constan_convolutions. Moved rgby filter under filters. Moved gaussian regulator into util.
