+++ slam_recognition/__init__.py
@@ -1,5 +1,5 @@
 __version__ = '0.0.2'
 
-from .center_surround_tensor import center_surround_tensor
-from .edge_tensor import stripe_tensor, simplex_stripe_tensors
-from . import zoom_tensor
+from slam_recognition.constant_convolutions.center_surround import center_surround_tensor
+from slam_recognition.constant_convolutions.edge_orientation_detector import stripe_tensor, simplex_stripe_tensors
+from slam_recognition.util import zoom
