+++ slam_recognition/constant_convolutions/center_surround/__init__.py
