+++ slam_recognition/constant_convolutions/edge_orientation_detector/__init__.py
