+++ slam_recognition/constant_convolutions/oriented_end_detector.py
