+++ slam_recognition/filters/__init__.py
@@ -1,1 +1,2 @@
-from .orientation_filter import orientation_filter
+from .orientation import orientation_filter
+from .rgby import rgby_filter
