+++ slam_recognition/filters/orientation.py
@@ -1,8 +1,9 @@
-from slam_recognition.rgby_filter import RGBYFilter
 import tensorflow as tf
 import numpy as np
-from slam_recognition.edge_tensor import rgb_2d_stripe_tensors
-from slam_recognition.regulator_tensor.gaussian_regulator_tensor import regulate_tensor, blur_tensor
+from slam_recognition.constant_convolutions.edge_orientation_detector import rgb_2d_stripe_tensors
+from slam_recognition.util.regulator import regulate_tensor
+from slam_recognition.constant_convolutions.gaussian_blur.gaussian_blur import blur_tensor
+from slam_recognition.util.get_dimensions import get_dimensions
 
 if False:
     from typing import Union
@@ -13,12 +14,7 @@ def orientation_filter(tensor,  # type: Union[tf.Tensor, np.ndarray]
                        ):
     """Mimics the blob cells in the lowest layer of the V1 in the neocortex, activating pixels that have high
             color difference."""
-    if isinstance(tensor, tf.Tensor):
-        dimensions = len(tensor.get_shape()) - 2
-    elif isinstance(tensor, np.ndarray):
-        dimensions = len(tensor.shape) - 2
-    else:
-        raise TypeError("Input to orientation filter must either be tensor or numpy array.")
+    dimensions = get_dimensions(tensor)
 
     simplex_boundaries_b = rgb_2d_stripe_tensors()  # todo: make n-d
     blur = blur_tensor(dimensions, lengths=blur_size)
