--- slam_recognition/regulator_tensor/__init__.py
@@ -1,1 +1,0 @@
-from .gaussian_regulator_tensor import blur_tensor, regulate_tensor
