+++ slam_recognition/util/regulator/__init__.py
@@ -1,0 +1,2 @@
+from .gaussian_regulator_tensor import regulate_tensor
+from slam_recognition.constant_convolutions.gaussian_blur.gaussian_blur import blur_tensor
