+++ slam_recognition/util/zoom/from_image.py
