+++ slam_recognition/util/zoom/to_image_list.py
